The floatingActionButton in the onboarding was previously an image. It removed and replaced with a dynamic button. Additionally, the images in the README.md file updated with new ones. Some bugs fixed, and minor improvements made.

+++ lib/pages/settings/settings_page_view_model.dart
@@ -10,8 +10,16 @@ class SettingsPageViewModel extends ChangeNotifier with BaseViewModel {
   final _navigationService = NavigationService.instance;
   final _authService = AuthService();
 
+  bool _isDeleteAccountLoading = false;
+
+  bool get isDeleteAccountLoading => _isDeleteAccountLoading;
+
   void deleteAccount(BuildContext buildContext) async {
+    if (_isDeleteAccountLoading) return;
+
     try {
+      _isDeleteAccountLoading = true;
+      notifyListeners();
       await _authService.deleteAccountPermanently();
       signOut(buildContext);
     } catch (e) {
@@ -20,6 +28,9 @@ class SettingsPageViewModel extends ChangeNotifier with BaseViewModel {
               e.toString()
           )
       );
+    } finally {
+      _isDeleteAccountLoading = false;
+      notifyListeners();
     }
   }
